+++ src/app/api/services/unit-role.service.ts
@@ -63,9 +63,20 @@ export class UnitRoleService extends CachedEntityService<UnitRole> {
         },
       },
       'observerOnly',
+      'mentorId',
+      'tutorNoteCount',
+      'canMarkOverflowTasks',
     );
 
-    this.mapping.addJsonKey('roleId', 'userId', 'unitId', 'role', 'observerOnly');
+    this.mapping.addJsonKey(
+      'roleId',
+      'userId',
+      'unitId',
+      'role',
+      'observerOnly',
+      'mentorId',
+      'canMarkOverflowTasks',
+    );
   }
 
   public createInstanceFrom(json: any, other?: any): UnitRole {