+++ src/app/tasks/modals/submission-type-modal/submission-type-modal.component.ts
@@ -13,12 +13,30 @@ export interface SubmissionTypeModalData {
   styleUrls: ['./submission-type-modal.component.scss'],
 })
 export class SubmissionTypeModalComponent {
+  selectedTransition: 'ready_for_feedback' | 'assess_in_portfolio' = null;
+
   constructor(
     @Inject(MAT_DIALOG_DATA) public data: SubmissionTypeModalData,
     private dialogRef: MatDialogRef<SubmissionTypeModalComponent>,
   ) {}
 
-  triggerTransition(status: TaskStatusEnum) {
+  public selectRff() {
+    this.selectedTransition = 'ready_for_feedback';
+  }
+
+  public selectAip() {
+    this.selectedTransition = 'assess_in_portfolio';
+  }
+
+  public submit() {
+    if (this.selectedTransition === null) {
+      return;
+    }
+
+    this.triggerTransition(this.selectedTransition);
+  }
+
+  private triggerTransition(status: TaskStatusEnum) {
     this.data.task.triggerTransition(status);
     this.dismissModal();
   }